requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiogram_dialog==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.16.3
annotated-types==0.7.0
APScheduler==3.11.0
asyncpg==0.30.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
coverage==7.10.1
dramatiq==1.18.0
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.4
packaging==25.0
pluggy==1.6.0
prometheus_client==0.22.1
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-cov==6.2.1
pytest-mock==3.14.1
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
redis==6.2.0
requests==2.32.4
setuptools==80.9.0
SQLAlchemy==2.0.41
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
urllib3==2.5.0
watchdog==6.0.0
yarl==1.20.1
zope.event==5.1
zope.interface==7.2
